refactor: flatten components folder
Since reorganizing our frontend and getting many of our basic components
from RIS UI, keeping some components inside "controls" doesn't make much
sense anymore. Moving all to the "components" folder.

RISDEV-7420
