+++ frontend/src/components/RisEmptyState.vue
