+++ frontend/src/components/RisNavbar.spec.ts
