+++ frontend/src/components/RisNavbarSide.spec.ts
