+++ frontend/src/views/amending-law/articles/editor/AmendingLawArticleEditor.view.vue
@@ -1,16 +1,17 @@
 <script setup lang="ts">
-import RisEmptyState from "@/components/controls/RisEmptyState.vue"
+import RisEmptyState from "@/components/RisEmptyState.vue"
 import RisLawPreview from "@/components/RisLawPreview.vue"
-import type { HeaderBreadcrumb } from "@/components/controls/RisHeader.vue"
-import RisHeader from "@/components/controls/RisHeader.vue"
-import RisLoadingSpinner from "@/components/controls/RisLoadingSpinner.vue"
+import type { HeaderBreadcrumb } from "@/components/RisHeader.vue"
+import RisHeader from "@/components/RisHeader.vue"
+import RisLoadingSpinner from "@/components/RisLoadingSpinner.vue"
 import RisCodeEditor from "@/components/editor/RisCodeEditor.vue"
-import RisTabs from "@/components/editor/RisTabs.vue"
+import RisTabs from "@/components/RisTabs.vue"
 import { useArticle } from "@/views/amending-law/articles/editor/useArticle"
 import { useEidPathParameter } from "@/composables/useEidPathParameter"
 import { useDokumentExpressionEliPathParameter } from "@/composables/useDokumentExpressionEliPathParameter"
 import { useAknElementEidSelection } from "@/views/amending-law/articles/editor/useAknElementEidSelection"
 import { useModHighlightClasses } from "@/composables/useTimeBoundaryHighlightClasses"
+import { useModHighlightClasses } from "@/composables/useZeitgrenzenHighlightClasses"
 import { useNormRenderHtml } from "@/composables/useNormRender"
 import { useNormXml } from "@/composables/useNormXml"
 import { getFrbrDisplayText } from "@/lib/frbr"
@@ -22,7 +23,7 @@ import type { LawElementIdentifier } from "@/types/lawElementIdentifier"
 import { useDebounce } from "@vueuse/core"
 import type { Ref } from "vue"
 import { computed, ref, toValue, watch } from "vue"
-import RisErrorCallout from "@/components/controls/RisErrorCallout.vue"
+import RisErrorCallout from "@/components/RisErrorCallout.vue"
 import { useRouter } from "vue-router"
 import { useModEidPathParameter } from "@/views/amending-law/articles/editor/useModEidPathParameter"
