+++ frontend/src/views/amending-law/articles/editor/multiple-mods/AmendingLawArticleEditorMultiMod.view.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
-import RisErrorCallout from "@/components/controls/RisErrorCallout.vue"
-import RisLoadingSpinner from "@/components/controls/RisLoadingSpinner.vue"
+import RisErrorCallout from "@/components/RisErrorCallout.vue"
+import RisLoadingSpinner from "@/components/RisLoadingSpinner.vue"
 import RisCodeEditor from "@/components/editor/RisCodeEditor.vue"
-import RisTabs from "@/components/editor/RisTabs.vue"
-import RisEmptyState from "@/components/controls/RisEmptyState.vue"
+import RisTabs from "@/components/RisTabs.vue"
+import RisEmptyState from "@/components/RisEmptyState.vue"
 import RisLawPreview from "@/components/RisLawPreview.vue"
 import { useDokumentExpressionEliPathParameter } from "@/composables/useDokumentExpressionEliPathParameter"
 import { useNormRenderHtml } from "@/composables/useNormRender"
