+++ frontend/src/views/amending-law/articles/editor/single-mods/AmendingLawArticleEditorSingleMod.view.vue
@@ -1,14 +1,14 @@
 <script setup lang="ts">
 import RisLawPreview from "@/components/RisLawPreview.vue"
 import RisModForm from "@/views/amending-law/articles/editor/single-mods/RisModForm.vue"
-import RisLoadingSpinner from "@/components/controls/RisLoadingSpinner.vue"
+import RisLoadingSpinner from "@/components/RisLoadingSpinner.vue"
 import RisCodeEditor from "@/components/editor/RisCodeEditor.vue"
-import RisTabs from "@/components/editor/RisTabs.vue"
+import RisTabs from "@/components/RisTabs.vue"
 import { useDokumentExpressionEliPathParameter } from "@/composables/useDokumentExpressionEliPathParameter"
 import { useNormRenderHtml } from "@/composables/useNormRender"
 import { useTemporalData } from "@/composables/useTemporalData"
 import { computed, ref } from "vue"
-import RisErrorCallout from "@/components/controls/RisErrorCallout.vue"
+import RisErrorCallout from "@/components/RisErrorCallout.vue"
 
 const xml = defineModel<string>("xml", {
   required: true,
