+++ frontend/src/views/expression/metadata-editor/ExpressionMetadataEditor.view.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import RisEmptyState from "@/components/controls/RisEmptyState.vue"
-import RisHeader from "@/components/controls/RisHeader.vue"
-import RisLoadingSpinner from "@/components/controls/RisLoadingSpinner.vue"
+import RisEmptyState from "@/components/RisEmptyState.vue"
+import RisHeader from "@/components/RisHeader.vue"
+import RisLoadingSpinner from "@/components/RisLoadingSpinner.vue"
 import { useDokumentExpressionEliPathParameter } from "@/composables/useDokumentExpressionEliPathParameter"
 import type { ComputedRef } from "vue"
 import { computed, ref, watch, nextTick } from "vue"
-import RisErrorCallout from "@/components/controls/RisErrorCallout.vue"
+import RisErrorCallout from "@/components/RisErrorCallout.vue"
 import Tree from "primevue/tree"
 import type { TreeNode } from "primevue/treenode"
 import ChevronUpIcon from "~icons/ic/baseline-keyboard-arrow-up"
