+++ frontend/src/views/expression/metadata-editor/element/ExpressionMetadataEditorElement.view.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import RisLawPreview from "@/components/RisLawPreview.vue"
-import RisErrorCallout from "@/components/controls/RisErrorCallout.vue"
-import { useHeaderContext } from "@/components/controls/RisHeader.vue"
-import RisLoadingSpinner from "@/components/controls/RisLoadingSpinner.vue"
+import RisErrorCallout from "@/components/RisErrorCallout.vue"
+import { useHeaderContext } from "@/components/RisHeader.vue"
+import RisLoadingSpinner from "@/components/RisLoadingSpinner.vue"
 import { useEidPathParameter } from "@/composables/useEidPathParameter"
 import { useElementId } from "@/composables/useElementId"
 import { useDokumentExpressionEliPathParameter } from "@/composables/useDokumentExpressionEliPathParameter"
@@ -77,8 +77,9 @@ const artNorm = computed<string | undefined>({
 })
 
 /* -------------------------------------------------- *
- * HTML preview                                 *
+ * HTML preview                                       *
  * -------------------------------------------------- */
+
 const {
   data: render,
   isFetching: renderIsLoading,
