+++ frontend/src/views/verkuendungen/verkuendungDetail/VerkuendungDetail.view.vue
@@ -1,10 +1,10 @@
 <script lang="ts" setup>
 import RisLawPreview from "@/components/RisLawPreview.vue"
 import RisViewLayout from "@/components/RisViewLayout.vue"
-import RisEmptyState from "@/components/controls/RisEmptyState.vue"
-import RisErrorCallout from "@/components/controls/RisErrorCallout.vue"
-import type { HeaderBreadcrumb } from "@/components/controls/RisHeader.vue"
-import RisLoadingSpinner from "@/components/controls/RisLoadingSpinner.vue"
+import RisEmptyState from "@/components/RisEmptyState.vue"
+import RisErrorCallout from "@/components/RisErrorCallout.vue"
+import type { HeaderBreadcrumb } from "@/components/RisHeader.vue"
+import RisLoadingSpinner from "@/components/RisLoadingSpinner.vue"
 import { useDokumentExpressionEliPathParameter } from "@/composables/useDokumentExpressionEliPathParameter"
 import { useElementId } from "@/composables/useElementId"
 import { getFrbrDisplayText } from "@/lib/frbr"
